[neutron] New manifest file to deploy Cisco N1KV-VEM
Background on 'N1KV-VEM’: It is the host-side component of N1KV
implementing the notion of Virtual Ports for VM Interfaces and also
handling the VM data-traffic. VEM is implemented as a couple of
user-space processes and also has a KLM for efficient data-path
switching.

VEM manifests does: setting up the Config-File for VEM service,
Starting VEM service and finally adding physical ports (along with
port-profile) to the VEM

Change-Id: Ibc8f1c4d2d743589e452ae6d1d87b5bfa123bc3a